fixed clang compile issue; build now automatically places pk3s in mod folder
The bundles made by Xcode need to be zipped and renamed before they can
be placed in a pk3, e.g. jampgame.bundle must be zipped and the zip
renamed to jampgame, which must then go in a pk3.

Now when you build this is all done automatically. A pk3 named after
the project is placed in the relevant mod folder (which should exist
already), e.g. jampgame.pk3. Set JKJAAPPSTORE in Build Settings to 1 if
you want the files to be placed for the App Store version of the game.
JKJANAME should be `Jedi Academy` for the Steam or App Store versions,
and `Jedi Academy MP` for other versions. JKJAMMODNAME should be your
mod's folder, so just `base` if it doesn't have its own folder. You'll
have to set these in all three projects.

clang was also complaining about inline stuff, so have done a temporary
fix

You can debug by setting the Executable correctly in
Product->Scheme->Edit Scheme.. for Run or by attaching to an already
running instance of JA.